pyproject.toml
[project]
name = "average-messenger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.129.0",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=9.0.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.46",
]
